[Ide] Fix memory leak when viewing code formatting options.
Opening the Solution Options dialog or the Project Options dialog and
selecting the Source Code - Code Formatting section would cause the
dialog and the options panels to not be removed from memory.

The CodeFormattingPanelWidget creates MimeTypePanelData objects which
reference MimeTypePolicyOptionsSection objects which reference the
options dialog. GLib.ManagedValue objects, from the Gtk.ListStore
keep the MimeTypePanelData objects alive which results in the dialog
still being referenced. Explicitly clearing the list store in the
CodeFormattingPanelWidget allows the objects to be garbage collected.