[Windows] Fix being unable to select a folder with SelectFolderDialog
The default constructor for the SelectFolderDialog has changed so it
now sets the default Action to be SelectFolder | CreateFolder. This
caused the select folder dialog to show the Open button instead of
the Select Folder button so only files could be selected. If the
SelectFolderDialog's Action was only set to SelectFolder then the
dialog would allow a folder to be selected. This fixes being
unable to select a folder when using the New Project dialog.