[SourceEditor] Added a file name check in the highlighting panel.
The operation failed silently. I think it's better to show an error
message than having a remove option because that gives the user time
to think about it.
A remove is clicked too fast and that operation isn't reversible - so
it's better to make that a bit more complex to do. (It's an unlikely
case anyways)