[NuGet] Do not keep reference to project removed from solution.
Fixed bug #38039 - [MemoryLeak] Unloading/reloading a project
https://bugzilla.xamarin.com/show_bug.cgi?id=38039

The ProjectTargetFrameworkMonitor, which is used to check if the NuGet
packages are compatible if the target framework changes, would keep a
reference to a project if it was removed from the solution or when
the project was reloaded when it was changed outside the IDE, which
causes the project to be removed and re-added. These references would
be cleared when the solution was closed but not whilst the solution
remained open. Now the NuGet addin does not keep a reference to a
project after it is removed from the solution.