[NuGet] Fix check for updates keeping project references.
Fixed bug #38043 - [MemoryLeak] NuGet check for updates
https://bugzilla.xamarin.com/show_bug.cgi?id=38043

Mono and debug builds on Windows keep a reference to the
CheckForUpdatesTask on the stack even though nothing is directly
using it. When the solution is closed this means that the projects
referenced by the CheckForUpdatesTask are kept in memory until
another solution is opened. Release builds on Windows seem to be
more aggressive in their garbage collection and will finalize the
CheckForUpdatesTask.

To prevent the CheckForUpdatesTask from keeping projects alive the
class has been changed so it does not keep any references to any
of the projects once it has finished checking or the check has been
cancelled, which can happen if the solution is closed before the
check has completed. This allows the projects to be garbage collected
when the solution is closed even if the CheckForUpdatesTask is not.