[NuGet] Fix package restore keeping reference to project.
With check for updates disable in preferences, but with automatic
package restore enabled, on opening and then closing the solution
the NuGet startup handler would indirectly keep a reference to the
project via the PackageRestorer.

When restoring packages for a single project the restore packages
in project handler would keep a reference to the project after the
solution was closed.