[build] Do not export `MONODEVELOP_PROFILE` by default
If the person working with XS wants to use an alternate profile, then
can export the env var themselves like they used to do for the last
few years.

This way `make run` is usable again as it does not use a temporary
profile which is blown away every time you clean the build.