[NuGet] Only refresh Packages and References folder in Solution window
Previously when any NuGet package was installed, removed or updated
the NuGet addin would refresh all child nodes of each project in the
Solution window. Now the NuGet addin only refreshes the Packages and
the References folder nodes since these are the only ones that need
to be updated by the NuGet addin.