[NuGet] Fix project referenced after NuGet install and solution closed
Fixed bug #38044 - [MemoryLeak] NuGet installing a package
https://bugzilla.xamarin.com/show_bug.cgi?id=38044

After installing, updating or removing a NuGet package and closing the
solution the project would still be referenced indirectly by the
BackgroundPackageActionRunner.

Also reduced the memory leak of the Add Packages dialog by clearing
the package items and removing references to the package view models
when the dialog is disposed. This prevents the project being
referenced after the solution is closed and also reduces the memory
leak due to the dialog remaining in memory. The memory leak for the
Add Packages dialog still exists.