[Ide] Fix References dialog keeping a reference to the project.
Fixed bug #38037 - [MemoryLeak] SelectReferenceDialog
https://bugzilla.xamarin.com/show_bug.cgi?id=38037

A single instance of the SelectReferencesDialog is used whilst the
IDE is running. If the dialog is opened and then the solution is
closed the dialog was keeping a reference to the project so it is not
garbage collected. Now when the dialog is closed the dialog and the
panels all remove all reference to any project and clear their lists
which allows the project to be garbage collected.