Notify file change event when saving a file
The NotifyFileChanged call was removed since in some cases the view already
notifies the change, and sending the event twice can be slow. The problem is
that we can't rely on the view to send the event, since not all views do it.
To reduce the impact of notificating the change event twice (when that
happens) we now do a freeze/thaw before and after the save operation,
so all file change events will be sent in a single event execution, so it should
be as fast as sending a single event.

Fixes bug #38100 - XS 6 Preview does not build project when there are changes