[NuGet] Fix incorrect update count displayed after updating packages
Fixed bug #37812 - Wrong solution tree label after updating packages
when package number decrease as part of update
https://bugzilla.xamarin.com/show_bug.cgi?id=37812

When an update caused a NuGet package to be uninstalled the Packages
folder could show an incorrect update available count since it still
counted the package that was uninstalled.

A way to reproduce this is to create an Android project, choosing
Max Compatibility. This installs 3 NuGet packages:

  package id="Xamarin.Android.Support.v13" version="20.0.0.4"
  package id="Xamarin.Android.Support.v4" version="20.0.0.4"
  package id="Xamarin.Android.Support.v7.AppCompat" version="20.0.0.2"

Updating all the NuGet packages in the project would result in the
Xamarin.Android.Support.v13 NuGet package being uninstalled but the
Packages folder would still show one update as being available. After
an update the NuGet packages installed would be:

  package id="Xamarin.Android.Support.v4" version="23.1.1.0"
  package id="Xamarin.Android.Support.v7.AppCompat" version="23.1.1.0"

Now the NuGet addin removes packages from its list of updated packages
if they are not in the packages.config file.