[Mac] Improve parsing of mime.types file.
Fixed bug #38672 - Can't open .m and .h files in the source code
editor
https://bugzilla.xamarin.com/show_bug.cgi?id=38672

Loading /etc/apache2/mime.types was not handling multiple file
extensions for a single mime type, nor handling file extensions
that included a number:

  video/x-m4v	m4v
  text/x-c	c cc cxx cpp h hh

This resulted in the file extension either being missing from the
mime map, when multiple file extensions were specified for a single
mime type in mime.types. Or it resulted in an incorrect file extension
being mapped to a mime type if the extension contained a number.