[GtkCore] Fix toolbox being empty when the GTK# designer is open.
The problem was that the ToolboxService was not finding the
IToolboxConsumer which is implemented on the DesignerPage since the
TabView was checking the Gtk.Widget casted to a Control so the
type check failed.

The Toolbox.xml file may need to be removed if the GTK# designer
was opened previously and no toolbox items were displayed so it
can be re-populated with toolbox items.