[NuGet] Show an error dialog if the NuGet.Config file cannot be read.
Previously if the NuGet.Config file could not be read the error would
be logged, but not reported to the user, and the NuGet addin would
then switch to using the default official NuGet package source
instead of those defined in the NuGet.Config file. Now an error
dialog is shown indicating that there was a problem reading the
NuGet.Config file.