[Core] Fix environment variables being removed from custom commands.
Fixed bug #1412 - Serializer keeps on discarding env vars from custom
commands
https://bugzilla.xamarin.com/show_bug.cgi?id=1412

When a change was made in the project options dialog any environment
variables defined for a custom command would be removed when the
project was saved. Other changes to a project, such as adding a file,
would not cause these variables to be removed.

  <CustomCommands>
    <Command type="Execute" command="notepad">
      <EnvironmentVariables>
        <Variable name="FOO" value="bar" />
      </EnvironmentVariables>
    </Command>
  </CustomCommands>

The problem was that when the project options dialog was opened the
configurations are cloned, which clones the custom commands, but the
environment variables were not cloned.