[Ide] Fix memory leak when adding new project to existing solution.
When adding a new project using the new project dialog the project
template would keep a reference to the parent solution folder that the
project was being added to. The project templates are not disposed
whilst the IDE is open so the solution and projects would remain
in memory after the solution was closed.

The project template was keeping the reference to the solution folder
through a ProjectCreateInformation object which was only being used
in a method which is no longer used so the method and the reference
to the ProjectCreateInformation have been removed from the
ProjectTemplate class which fixes the memory leak.