[Ide] Fix hint path not resolved for new F# PCL project.
Fixed bug #39111 - fsharp.core not added to a 2nd fsharp PCL project
in a solution
https://bugzilla.xamarin.com/show_bug.cgi?id=39111

When a new F# PCL project is added to an existing solution the
FSharp.Core reference in the Solution window shows an error
indicating that the assembly is not available for the PCL project.
This does not happen when creating a F# PCL project without an
existing solution.

The problem is that the ProjectReference does not resolve the
hint path for the FSharp.Core assembly when the project is added to
an existing solution. The F# project template defines a hint path
of:

    <Reference Include="FSharp.Core">
      <HintPath>$(MSBuildExtensionsPath32)\..\Reference Assemblies\Microsoft\FSharp\.NETCore\$(TargetFSharpCoreVersion)\FSharp.Core.dll</HintPath>
    </Reference>

The ProjectReference holds two hint paths, one is the original
MSBuild hint path and the other is the evaluated/resolved path.
When loading the project from a file or opening a solution containing
a F# PCL project, the MSBuild hint path is evaluated and the correct
assembly is found. When a new F# PCL project is added to an existing
solution the hint path is not resolved and remains null. The
FSharp.Core reference is then resolved incorrectly to the assembly
in the GAC.

One way to fix this is to reload the project so the hint path is
correctly resolved. This is implemented in this commit. Note that the
case where a new solution containing a F# PCL project is created and
added to a workspace is also handled. In this case all the projects
in the solution need to be reloaded. When creating a new solution
without a parent workspace there is no need to reload the projects
since they are reloaded when the new solution is opened which happens
after everything is created and saved.

Another way to fix this is to resolve the hint path from the
MSBuild hint path when the project is created by the New Project
dialog and pass this resolved hint path to the ProjectReference.