[Refactoring] Ensure the JumpToDeclarationHandler extension chain is queried when navigating to a source symbol.
Previously the code had a few short-circuit and optimizations that prevented the JumpToDeclarationHandler extension chain from being queried and instead tried to convert the caret into a Roslyn symbol early on that could be directly jumped to.

This is an issue if you want to provide a custom jump handler that override the default navigation behavior of a valid source symbol.

For instance, with Android projects it's useful to replace the default navigation strategy of `Resource.*.*` fields to jump to the corresponding resource file instead of opening the autogenerated Resource.designer.cs class.