[Core] Fix project serialization issues
When a property is removed from the main property group, make sure it is
also removed from the evaluated properties list.

Fixed serialization issue when an item with wildcards has metadata that
references a property.

Fixes bug #38940 - Switching XM target framework from Mobile to 4.5
with a PCL reference leaves project as unloadable