[NuGet] Fix spacing used for multiple packages in license dialog.
When installing a NuGet package such as Microsoft.AspNet.Mvc the
license dialog shows three packages with licenses that need to be
accepted. Each package shown had too much space around it since it
was filling all the available space. Now the package information
uses only the space it needs.