[Refactoring] Add dummy MidStep in JumpToDeclarationHandler chain to make it easier for extension to position themselves.
This is amending for commit a12b620681f and based on comment https://github.com/mono/monodevelop/commit/a12b620681fc6ce253e56c524d40438547f4946f#commitcomment-16488927

Instead of relying on some id value, we initialize the extension chain with a placeholder the does nothing and set `defaultInsertAfter` on the extension point so that by default new extensions are added after it. It then becomes easy for extensions that need to override default behavior to specify instead a `insertBefore` attribute so that they are queried first.