[NUnit3] Fix runner issue
For NUnit 2 we collect and provide all tests to be executed to the runner.
That doesn't work well on NUnit 3. In that case we can just provide the
root test suite, and it will just execute all included tests.

Fixes bug #39259 - [NUnit3] Xamarin Studio Nunit3 runner does not
respect ExplicitAttribute