Branding service improvements
The BrandingService's ApplicationName is no longer readonly. If the
UpdateApplicationName method is used to update the name then the
ApplicationNameChanged event will fire. This can be used to update
any part of the UI that is showing the old application name.

Moved the default name of the dialog image to a class so it can
be updated at runtime.