[Ide] Fix error dialog not shown on Windows if GTK# is not installed
With GTK# not installed the GTK# check on Windows was not reached.
This resulted in a silent failure and no error dialog was displayed.
When the IdeApp's constructor is initialized it tries to use the
FontService which throws an exception since it tries to use GTK#
before the check for GTK# on Windows is run.

System.TypeInitializationException: The type initializer for 'MonoDevelop.Ide.Fonts.FontService' threw an exception. ---> System.DllNotFoundException: Unable to load DLL 'libpango-1.0-0.dll': The specified module could not be found. (Exception from HRESULT: 0x8007007E)
      at Pango.FontDescription.pango_font_description_new()
         at Pango.FontDescription..ctor() in c:\MinGW\msys\1.0\home\builder\gtk-sharp\pango\generated\FontDescription.cs:line 291
            at MonoDevelop.Ide.Fonts.FontService..cctor() in src\core\MonoDevelop.Ide\MonoDevelop.Ide.Fonts\FontService.cs:line 43
               --- End of inner exception stack trace ---
               at MonoDevelop.Ide.Fonts.FontService.GetFontProperty(String name)
                  at MonoDevelop.Ide.IdePreferences..ctor() in src\core\MonoDevelop.Ide\MonoDevelop.Ide\IdePreferences.cs:line 129
                  at MonoDevelop.Ide.IdeApp..cctor() in src\\core\\MonoDevelop.Ide\\MonoDevelop.Ide\\Ide.cs:line 110

   at MonoDevelop.Ide.IdeApp.set_Customizer(IdeCustomizer value)
   at MonoDevelop.Ide.IdeStartup.Run(MonoDevelopOptions options) in src\core\MonoDevelop.Ide\MonoDevelop.Ide\IdeStartup.cs:line 85
   at MonoDevelop.Ide.IdeStartup.Main(String[] args, IdeCustomizer customizer) in src\core\MonoDevelop.Ide\MonoDevelop.Ide\IdeStartup.cs:line 593

To fix this the first call to IdeApp has now been moved after the
GTK# check on Windows. Now if GTK# is not installed on Windows an
error dialog is now displayed indicating that GTK# is missing and can
open the appropriate web page with the GTK# installer download.