Fix content destroy issue
When a view has attached views, and the view that is shown by default is not
the main view, iit may happen that the main view is not added to the notebook
and in that case it won't be destroyed when the document is closed.

This fix checks if the main view has a parent. If it doesn't it means that it has
not been added to the notebook, so it has to be explicitly destroyed when
closing the document.

Fixes bug #39591 - Moving folder between projects fails to add files.