[NuGet] Insert MSBuild property files ( .props) at start of project.
Fixed bug #19054 - NuGet package MSBuild property files (.props)
should be added to the start of the project file

Installing a NuGet package that included an MSBuild .props file would
add an Import element for the .props at the end of the project file
(.csproj) which is incorrect. Now .props files are added to the
project file as the first child element inside the Project's root
element.