[Core] Fixed import not being added to project.
Fixed bug #39766 - Uninstalling and then installing a NuGet package
does not add back imports
https://bugzilla.xamarin.com/show_bug.cgi?id=39766

After uninstalling a NuGet package that used a custom MSBuild .targets
file, such as Xamarin.Forms, when the same package version was
installed again the Import element for the NuGet package's MSBuild
.targets file was not saved in the project file. This was because
every import added or removed from the project was not cleared after
the project was saved. So the import would be removed again every time
the project was saved until the solution was closed and re-opened
again.