[Core] Fix new line inserted when adding import at top of project.
Inserting a new Import as the first child in the MSBuildProject
using AddNewImport was inserting a blank new line before the Import
element. This was because when an MSBuild project file is loaded
and contains the xml declaration line the Project element's
start whitespace is a newline. When inserting a child element as the
first child of the project the start whitespace is taken from the
parent, which in this case is a newline, and then an extra two spaces
are added. This gives the Import element an incorrect start whitespace
since it contains the newline when it should just contain the spaces.