[Ide] Fix misleading label in New Project dialog.
Fixed bug #39886 - Checkbox text has meaning opposite to its function
https://bugzilla.xamarin.com/show_bug.cgi?id=39886

The New Project dialog had a check box with the label "Create a
project within the solution directory". When this is checked a
directory is created for the project inside the solution directory
however the label suggests that the project will be created in
the solution directory.

The label has been changed to "Create a project directory within the
solution directory" since this is a better description of what will
occur if this check box is checked.