[Ide] Fix ErrorListPad button icons
The internal Gtk logic to align and show the label and
the icon of a ToggleButton seems to break at some
point when the label is updated too fast and too often.
This workaround puts the text and the icon in a box
and adds them to the buttons, overriding the internal
GtkButton child handling. When the label text is being
updated, the button will not rearrange its children now.

(fixes bug #38810)

(cherry picked from commit 2743faf4bb8a33563abe062df40e286dcac212ed)