[Ide] Make opening PackageReferencePanel faster.
Adds a Freeze/Thaw child notify and modifies appends to bei nserts.

This optimizes the launching from 1.5s to 0.5s on my machine.

Gtk docs say that InsertWithValues is definitely faster than AppendValues.

But gtk# also seems to be doing something wrong as it's not using the vararg method to set a value, but sets each column separately.