[Ide] Always complete the task when running CustomToolService.Update
If we get 5 updates one right after the other, we will hit the case
where cancellation has been requested while we are waiting for the
previous task to complete.

Assume thread 1 is inside `await existingTask`. Thread 2 pops in,
notices that there's an existing task for the file it's working on,
cancels it and then places a new object in the `runningTasks` dictionary.
Thread 2 now awaits the task associated with the old object in
the `runningTasks` dictionary.

Now thread 1 completes it's await and sees it has been cancelled. If
it does not mark it's `task` as complete then thread 2 will be stuck
forever.