[NuGet] Add missing information to Add Packages dialog.
Project url, license url and published date now displayed.
Published date is not being returned for the NuGet v3 package source.
Dependencies are hidden since it is not being returned for the
NuGet v3 package source.

Use package title instead of id if it is available. For example,
Newtonsoft.Json has a package title of Json.NET and a package id
of Newtonsoft.Json, so the Json.NET title is displayed as the name.