[NuGet] Fix error installing different package version into a project
When installing a NuGet package into a project that had the same NuGet
package installed but a different version the install would fail. When
uninstalling the existing NuGet package the NuGet package manager
looks at all projects in the solution to see if they use that NuGet
package. This was failing since the solution manager did not support
creating NuGetProjects for all projects in the solution.