[NuGet] Restore packages before updating.
When updating all NuGet packages in a project then missing NuGet
packages for that project are restored before the update.
When updating all NuGet packages in a solution then missing NuGet
packages are restored before the update.

When updating a single NuGet package then a restore is not done before
the update. NuGet 3 handles this and will successfully update the
NuGet package if it is missing. If there are no updates then the
NuGet package is not restored.