[NuGet] Add package version combo box to Add Packages dialog.
Can now add a specific NuGet package version by selecting the
version in the combo box. The version combo box is displayed at
the bottom of the package information on the right hand side of
the dialog. The package version information is read when a NuGet
package is selected in the dialog.