[NuGet] Fix solution window not updating when last package removed.
When the last NuGet package was removed from the project the Solution
window would still show the NuGet package in the Packages folder and
the From Packages folder would still be displayed with no children.

When the packages.config file is removed from the project during an
uninstall the solution window is now refreshed.