[Ide] Restrict where the edition name displayed is used.
Fixed bug #40133 - Edition shown in places it shouldn't be
https://bugzilla.xamarin.com/show_bug.cgi?id=40133

The edition was added to the application name globally so it affected
all parts of the IDE. Now the application's edition is only shown in
the About dialog, the status area and the main window title on
Windows.

The BrandingService now has two application names - a short IDE name
(ApplicationName) and one that includes the edition
(ApplicationLongName).

The BrandingService has also been changed so that these two
application names are properties that can be set instead of fields so
the separate setter method can be removed.