[NuGet] Fix Packages folder not updating with project.json changes.
When a NuGet package was installed or removed in a project using a
project.json file the Packages folder was not updated. This is
because no file changed events occur for the project.json file since
all changes are done by NuGet 3 and not by the NuGet addin.

Now when the PackageOperationsFinished event is used to trigger a
refresh of the package information. This event fires when the current
set of installs, update or removes are all processed by the
background package action runner. Using this event also means that
the change added in commit bb747ebdfcca05a25b0fbc1ebcf389348b103d80 is
no longer required so that change has been reverted.