[NuGet] Fix wrong NuGet package version being installed.
If a package version is not specified then using the aggregate
repository with the machine cache as the priority repository will
use the highest version found in the local machine cache instead
of the aggregate repository. This can result in a lower version being
installed than what is available on a remote package source.

One example of this problem is when the Google Play Services dialog
is used. This installs a Google Play Service NuGet package but does
not specify a version. If the NuGet package being installed is
found in the local machine cache then it will be used instead of
the latest from nuget.org.