[NuGet] Fix update information not removed from Packages folder.
The available NuGet package updates shown in the Packages folder was
not updated when the packages were updated. Previously when a NuGet
package was installed or updated an event was fired with the package
and all the operations which was used to update the list of updated
NuGet packages.

A different approach has now been taken where the updated packages
information is updated when thePackages folder is refreshed based on
the currently installed packages in the project instead of relying
on an event.