[NuGet] Make adding local package sources easier in preferences.
Fixed bug #28665 - Package sources manager should have browse button
https://bugzilla.xamarin.com/show_bug.cgi?id=28665

There is now a browse button available when adding or editing a
package source in the preferences dialog which can be used to select a
folder on the local machine. Also the dialog has been changed to
indicate that a url or a folder can be used. Previously it was not
obvious that a folder could be used as a package source.