[NuGet] Fix update information not shown on some projects.
When a solution contained multiple projects sometimes the updated
package information was not displayed for all projects. The problem
was that the updated package information was being cleared before
the installed packages were read from the project. This was caused by
the change made in commit f777a37864ff9a3e67ad22b2f83f8199674134d5

Now the Packages folder does not read or clear the updated package
information if it has not first read any NuGet package reference
information about the project.