[NuGet] Fix From Packages being shown with no references.
The From Packages folder was shown inside the References when the
project had a package reference even if there were no assemblies being
referenced from the packages directory. Now the From Packages folder
is not added to the References folder when there are no assemblies
referenced from the packages directory.