[NuGet] Fix license acceptance dialog not being displayed.
The license acceptance dialog was not being displayed when installing
or updating a NuGet package that requires the license to be accepted.
Note that the behaviour of this feature has changed. Previously the
dialog was only shown if the same NuGet package was not already
installed into any project. Currently the license acceptance dialog
is always shown when a NuGet package is installed into a project even
if it was previously installed into the solution. This is the same
behaviour as NuGet 3 has in Visual Studio 2015. However this may be
changed back to the old behaviour.