[MacPlatform] Fix CommandArray menu item leak
When the items of an CommandArray menu entry
are being updated, we first remove all previously
inserted items. The cleanup routine expects
them to be directly after the hidden target command
item, but this was not the case because PopulateArrayItems
inserts the items at the given index (which
was the position of the target itself and not
the position after/behind it), thus the old (hidden)
items were not detected and removed correctly from
the parent menu.

(cherry picked from commit 135233c3af0fa0516109a599b7b5346f57039715)