[MacPlatform] Fix CommandArray menu item leak
When the items of an CommandArray menu entry
are being updated, we first remove all previously
inserted items. The cleanup routine expects
them to be directly after the hidden target command
item, but this was not the case because the index has
beend increased after each new item was added, so
the item was inserted before and not after the
target command (index always lower then the target),
thus the old (hidden) items were not detected and
removed correctly from the parent menu.