[NuGet] Restore failures now reported.
Failure to restore NuGet packages for projects using a packages.config
file were not reported. NuGet's PackageRestoreManager has a
PackageRestoreFailedEvent which is fired when a failure occurs during
a restore. This event is now monitored and failures are logged to the
Package Console and after the restore completes the failure is
reported in the status area.