[Core] Fix invalid configuration mapping for new generic projects.
Fixed bug #40507 - On creating Generic project, it is showing error
in Solution pad.
https://bugzilla.xamarin.com/show_bug.cgi?id=40507

Creating a new generic project would show an error icon next to the
project in the Solution pad. The error message was 'Invalid
configuration mapping'. The generic project was created without any
configuration property group since the properties inside this group
had the default values so the entire group is removed when the project
was saved.

  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Default|AnyCPU' ">

Now the OutputPath uses a different default value so the configuration
property group is saved in the generic project. The generic project
file created now matches what is created with Xamarin Studio 5.10.