[Ide] Signal that the last task was removed to progressmonitor users.
This adds a LastTaskRemoved entrypoint where subscribers can decide
whether to end animations relating to the current progress of a monitor.

The current ProgressMonitor behaviour is to replace the rootTask if none
exists.

In the case of the Rebuild command, a Clean Task would be attached, then
finished, afterwards a Build Task would be attached, then finished.

The Clean Task would end up with a Progress > 0, so the animation would
progress in the statusbar.

The newly placed Build Task would end up with a Progress = 0, in a context
where a progress is started, but never ended. Therefore the progressbar
would animate backwards to 0 and continue forward with the Build Task's
result.

Bug 39314 - Progress bar goes backwards when building solution