[NuGet] Use NuGet 3 when installing packages from components.
The public API provided by the NuGet addin now uses NuGet 3 when
installing NuGet packages. This is currently only used when a
Xamarin Component installs a NuGet package.

Add another install method that allows the caller to specify whether
the NuGet package licenses have already been accepted or not. This
can be used to prevent the license acceptance dialog from being
displayed. The default behaviour is to show the license acceptance
dialog.