[NuGet] ProjectOperations now finds installed packages using NuGet 3.
The installed NuGet packages can be retrieved through the public API
the NuGet addin provides. This has now been changed to use NuGet 3
and the NuGetProject's GetInstalledPackages. If this fails then the
AggregateException is unwrapped and the underlying exception is thrown
so a better error is available for the caller than 'One or more errors
occurred.'.

Also handle the PackageManagementWorkspace not having a solution when
it is asked for a solution manager. This can happen when a component
is bound to a project before the Ide's SolutionLoaded event is fired.